Fix x86_64 declaration of uint64_t.

+++ sortix/include/sortix/x64/bits.h
@@ -39,13 +39,13 @@
 typedef signed char __int8_t;
 typedef signed short __int16_t;
 typedef signed int __int32_t;
-typedef signed long long __int64_t;
+typedef signed long __int64_t;
 
 /* Define basic unsigned types. */
 typedef unsigned char __uint8_t;
 typedef unsigned short __uint16_t;
 typedef unsigned int __uint32_t;
-typedef unsigned long long __uint64_t;
+typedef unsigned long __uint64_t;
 
 typedef unsigned int __nat;
